+++ src/Linkification/Embedding.js
@@ -592,7 +592,7 @@ var Embedding = {
     }
     , {
       key: 'Twitter',
-      regExp: /^\w+:\/\/(?:www\.|mobile\.)?twitter\.com\/(\w+\/status\/\d+)/,
+      regExp: /^\w+:\/\/(?:www\.|mobile\.)?(?:twitter|x)\.com\/(\w+\/status\/\d+)/,
       style: 'border: none; width: 550px; height: 250px; overflow: hidden; resize: both;',
       el(a) {
         const el = $.el('iframe');